\name{COZIGAM-package}
\alias{COZIGAM-package}
\alias{COZIGAM}
\docType{package}
\title{Constrained and Unconstrained Zero-Inflated Generalized
   Additive Models with Model Selection Criterion}
\description{
Constrained and Unconstrained Zero-Inflated Generalized
Additive Models (ZIGAM) fitting with associated model
plotting, prediction and selection.}
\details{
\tabular{ll}{
Package: \tab COZIGAM\cr
Type: \tab Package\cr
Version: \tab 2.0-2\cr
Date: \tab 2009-7-4\cr
License:\tab GPL version 2 or newer\cr
}
\tabular{ll}{
COZIGAM.cts  \tab       Fitting Component-Specific-Proportionally
                        Constrained Zero-Inflated Generalized Additive
                        Models with Continuous Responses \cr
COZIGAM.dis  \tab       Fitting Component-Specific-Proportionally
                        Constrained Zero-Inflated Generalized Additive
                        Models with Discrete Responses   \cr
PCOZIGAM.cts \tab       Fitting Proportionally Constrained
                        Zero-Inflated Generalized Additive Models with
                        Continuous Responses   \cr
PCOZIGAM.dis \tab       Fitting Proportionally Constrained
                        Zero-Inflated Generalized Additive Models with
                        Discrete Responses   \cr
ZIGAM.cts    \tab       Fitting Continuous Zero-Inflated Generalized
                        Additive Models   \cr
ZIGAM.dis    \tab       Fitting Discrete Zero-Inflated Generalized
                        Additive Models   \cr
alaska       \tab       Alaska Peninsula Coastline \cr
cozigam      \tab       Fitting Constrained Zero-Inflated Generalized
                        Additive Models   \cr
disgam       \tab       Fitting Discrete Generalized Additive Models
                        with Model Selection Criterion    \cr
eggdata      \tab       Pollock Egg Data     \cr
plot.cozigam \tab       Default COZIGAM plotting   \cr
predict.cozigam \tab    Prediction from fitted COZIGAM \cr
print.cozigam   \tab    Constrained Zero-Inflated Generalized Additive
                        Model default print statement \cr
summary.cozigam \tab    Summary for a COZIGAM fit  \cr
test         \tab       Some Test Functions used in Simulation Study  \cr
zigam        \tab       Fitting (Unconstrained) Zero-Inflated Generalized Additive Models   \cr
}
}
\author{Hai Liu and Kung-Sik Chan;
Maintainer: Hai Liu <hailau@gmail.com>
}
\keyword{ package }
