\name{eggdata}
\docType{data}
\alias{eggdata}
\title{Pollock Egg Data}
\description{
This data set gives Pollock egg count data from cruise studies off the tip of Alaska in year of 1987.
}
\author{
Hai Liu and Kung-Sik Chan
}
\usage{data(eggdata)}
\format{A data frame with 274 observations on the following 6 variables.
\describe{
   \item{\code{bottom}}{A numeric vector: bottom depth at the sampling spot}
   \item{\code{lat}}{A numeric vector: latitude of the sampling spot}
   \item{\code{lon}}{A numeric vector: longitude of the sampling spot}
   \item{\code{catch}}{A numeric vector: the number of eggs caught in the unit of CPUE (Capture Per Unit Effort) and ``0" indicates a zero catch}
   \item{\code{j.day}}{Julian day}
   \item{\code{year}}{Year: 1987}
}
}
\keyword{datasets}
