\name{print.cozigam}
\alias{print.cozigam}
\title{Constrained Zero-Inflated Generalized Additive Model default print
statement}
\description{
This is the default print function for a COZIGAM object.
}
\author{
Hai Liu and Kung-Sik Chan
}

\seealso{
   \code{\link{cozigam}},\code{\link{summary.cozigam}}
}
\keyword{smooth} \keyword{models} \keyword{regression}

