\name{test}
\alias{test}
\alias{f0}
\alias{f1}
\title{Some Test Functions used in Simulation Study}
\description{
\code{test} defines a 2-D test function.

\code{f0} and \code{f1} define two 1-D test functions. The three test
functions will be used in the simulation study of COZIGAM fits.
}
\author{
Hai Liu and Kung-Sik Chan
}

\references{
Liu, H and Chan, K.S. (2008) Constrained Generalized Additive Model with Zero-Inflated Data. 
}
\seealso{
   \code{\link{cozigam}}
}
\keyword{smooth} \keyword{models} \keyword{regression}

