\name{ZIGAM.cts}
\alias{ZIGAM.cts}
\title{Fitting Continuous Zero-Inflated Generalized Additive Models}
\description{
\code{ZIGAM.cts} fits a ZIGAM for continuous zero-inflated exponential family responses. 

\code{\link{ZIGAM.dis}} fits a ZIGAM for discrete zero-inflated exponential family 
responses. One of the two funcions is called from \code{\link{zigam}}
automatically depending on the distribution family. 
}
\author{
Hai Liu and Kung-Sik Chan
}

\references{
Liu, H and Chan, K.S. (2008) Constrained Generalized Additive Model with Zero-Inflated Data. 
Technical Report 388, Department of Statisics and Actuarial Science, The Unversity of Iowa. 
\url{http://www.stat.uiowa.edu/techrep/tr388.pdf} 
}
\seealso{
   \code{\link{zigam}}
}
\keyword{smooth} \keyword{models} \keyword{regression}


