\name{COZIGAM.cts}
\alias{COZIGAM.cts}
\alias{COZIGAM.dis}
\title{Fitting Constrained Zero-Inflated Generalized Additive Models with
Continuous Response}
\description{
\code{COZIGAM.cts} fits a COZIGAM with continuous exponential family 
responses. 

\code{COZIGAM.dis} fits a COZIGAM with discrete exponential family 
responses. One of the two funcions is called from \code{\link{cozigam}} 
automatically depending on the distribution family used. 
}
\author{
Hai Liu and Kung-Sik Chan
}

\references{
Liu, H and Chan, K.S. (2008) Constrained Generalized Additive Model with Zero-Inflated Data. 
}
\seealso{
   \code{\link{cozigam}},\code{\link{summary.cozigam}}
}
\keyword{smooth} \keyword{models} \keyword{regression}


