\name{eggdata}
\docType{data}
\alias{eggdata}
\title{Pollock Egg Data}
\description{
This data set gives Pollock egg count data from cruise studies off the tip of Alaska in early part of 1986.
}
\author{
Hai Liu and Kung-Sik Chan
}
\usage{data(eggdata)}
\format{A data frame with 337 observations on the following 6 variables.
   \item{bottom}{A numeric vector: bottom depth at the sampling spot.}
   \item{lat}{A numeric vector: latitude of the sampling spot.}
   \item{lon}{A numeric vector: longitude of the sampling spot.}
   \item{catch}{A numeric vector: the number of eggs caught in the unit of CPUE (Capture Per Unit Effort)
   and "0" indicates a zero catch.}
   \item{j.day}{Julian day.}
   \item{year}{Year number: 1986.}
}
\keyword{datasets}