\name{print.cozigam}
\alias{print.cozigam}
\title{Constrained Zero-Inflated Generalized Additive Model default print
statement}
\description{
This is the default print statement for a COZIGAM object.
}
\author{
Hai Liu and Kung-Sik Chan
}

\references{
Liu, H and Chan, K.S. (2008) Constrained Generalized Additive Model with Zero-Inflated Data. 
}
\seealso{
   \code{\link{cozigam}},\code{\link{summary.cozigam}}
}
\keyword{smooth} \keyword{models} \keyword{regression}

