\name{ZIGAM.cts}
\alias{ZIGAM.cts}
\title{Fitting Continuous Zero-Inflated Generalized Additive Models}
\description{
\code{ZIGAM.cts} fits a ZIGAM for continuous zero-inflated exponential family responses. 

\code{\link{ZIGAM.dis}} fits a ZIGAM for discrete zero-inflated exponential family 
responses. One of the two funcions is called from \code{\link{zigam}}
automatically depending on the distribution family. 
}
\author{
Hai Liu and Kung-Sik Chan
}

\references{
Liu, H and Chan, K.S. (2010)  Introducing COZIGAM: An R Package for Unconstrained and Constrained Zero-Inflated Generalized Additive Model Analysis. Journal of Statistical Software, 35(11), 1-26.
\url{http://www.jstatsoft.org/v35/i11/}
}
\seealso{
   \code{\link{zigam}}
}
\keyword{smooth} \keyword{models} \keyword{regression}


